+++ ckanext/dcat/tests/test_euro_dcatap_profile_parse.py
@@ -321,7 +321,7 @@ class TestEuroDCATAPProfileParsing(BaseParseTest):
         extras = self._extras(dataset)
         assert extras['contact_email'] == 'contact@some.org'
 
-    def test_dataset_access_rights_and_distribution_rights_rights_statement(self):
+    def test_dataset_access_rights_and_distribution_rights_rights_statement_literal(self):
         # license_id retrieved from the URI of dcat:license object
         g = Graph()
 
@@ -351,6 +351,36 @@ class TestEuroDCATAPProfileParsing(BaseParseTest):
         assert extras['access_rights'] == 'public dataset'
         resource = dataset['resources'][0]
         assert resource['rights'] == 'public distribution'
+
+    def test_dataset_access_rights_and_distribution_rights_rights_statement_uriref(self):
+        g = Graph()
+
+        dataset_ref = URIRef("http://example.org/datasets/1")
+        g.add((dataset_ref, RDF.type, DCAT.Dataset))
+
+        # access_rights
+        access_rights = BNode()
+        g.add((access_rights, RDF.type, DCT.RightsStatement))
+        g.add((access_rights, RDFS.label, URIRef("http://example.org/datasets/1/ds/3")))
+        g.add((dataset_ref, DCT.accessRights, access_rights))
+        # rights
+        rights = BNode()
+        g.add((rights, RDF.type, DCT.RightsStatement))
+        g.add((rights, RDFS.label, URIRef("http://example.org/datasets/1/ds/2")))
+        distribution = URIRef("http://example.org/datasets/1/ds/1")
+        g.add((dataset_ref, DCAT.distribution, distribution))
+        g.add((distribution, RDF.type, DCAT.Distribution))
+        g.add((distribution, DCT.rights, rights))
+
+        p = RDFParser(profiles=['euro_dcat_ap'])
+
+        p.g = g
+
+        dataset = [d for d in p.datasets()][0]
+        extras = self._extras(dataset)
+        assert extras['access_rights'] == 'http://example.org/datasets/1/ds/3'
+        resource = dataset['resources'][0]
+        assert resource['rights'] == 'http://example.org/datasets/1/ds/2'
 
     def test_distribution_access_url(self):
         g = Graph()